Fix set_loop_opt legality (closes #111)
Added a `legal` member to the `Scop` class which changes according to
the legality of the state the `scop` is in.

Changed `set_loop_opt` transformation to return -1 (instead of 1),
which is means "scop.legal" stays the same.

Added docs documenting the `pool_idx`, `scop_idx` and return values in
`ctadashi.cc`.